V.0.4 librerias
