+++ nbproject/project.properties
@@ -76,7 +76,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=Mortycio
-main.class=contenido.Login
+main.class=View.Login
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
