+++ src/View/LoginRegistro.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
  */
-package contenido;
+package View;
 
 import java.awt.Color;
 import javax.swing.JOptionPane;
